refactor: Int 타입을 date 포맷의 String으로 변환하는 함수 분리
* 다른 VC에서도 사용할 때를 생각해서
  Int의 extension으로 changeDateFormat() 메서드를 분리하여 재사용 하도록 수정
